kernel: [IPV6]: Convert cork.hop_limit and cork.tclass into u8 instead of int.

+++ linux-2.6.21.x/include/linux/ipv6.h
@@ -326,8 +326,8 @@ struct ipv6_pinfo {
 	struct {
 		struct ipv6_txoptions *opt;
 		struct rt6_info	*rt;
-		int hop_limit;
-		int tclass;
+		u8 hop_limit;
+		u8 tclass;
 	} cork;
 };
